fix pattern eval and change brush from step only to linear only (see below)
So far, all the properties on brushes can be modified on painter nodes, so they can be linear when rendered using hapev2.
Parameter changes in the primitive equation have to be "step" to render using hapev2 (AFAIK).
Streaming API has no such restrictions, but im restricting it to hapev2 features for now.